Move ImageUtil methods that depend on metadata back to iped-viewers.
